Renamed Crossref providers module names
