+++ sciety_labs/providers/opensearch/async_providers.py
@@ -15,7 +15,7 @@ from sciety_labs.providers.interfaces.article_recommendation import (
 from sciety_labs.providers.interfaces.async_article_recommendation import (
     AsyncSingleArticleRecommendationProvider
 )
-from sciety_labs.providers.crossref.async_crossref import AsyncCrossrefMetaDataProvider
+from sciety_labs.providers.crossref.async_providers import AsyncCrossrefMetaDataProvider
 from sciety_labs.providers.semantic_scholar.async_providers import (
     AsyncSemanticScholarTitleAbstractEmbeddingVectorProvider
 )
